Flytt sanity prosjekt inn i selbetjening-ui
