--- apps/brukerdialog-sanity/sanity.cli.ts
@@ -1,8 +1,0 @@
-import {defineCliConfig} from 'sanity/cli'
-
-export default defineCliConfig({
-  api: {
-    projectId: 'u0dlg8d8',
-    dataset: 'production'
-  }
-})
