+++ apps/selvbetjening-ui/sanity/.eslintrc
