+++ apps/selvbetjening-ui/sanity/.gitignore
