+++ apps/selvbetjening-ui/sanity/README.md
@@ -1,0 +1,23 @@
+# Brukerdialog Sanity
+
+## Logg inn med SSO
+
+`npx sanity login --sso navikt`
+
+## Deploy til Sanity
+
+`yarn deploy`
+
+## Lokal kjøring
+
+`yarn dev`
+
+# Sanity Clean Content Studio
+
+Congratulations, you have now installed the Sanity Content Studio, an open-source real-time content editing environment connected to the Sanity backend.
+
+Now you can do the following things:
+
+-   [Read “getting started” in the docs](https://www.sanity.io/docs/introduction/getting-started?utm_source=readme)
+-   [Join the community Slack](https://slack.sanity.io/?utm_source=readme)
+-   [Extend and build plugins](https://www.sanity.io/docs/content-studio/extending?utm_source=readme)
