+++ apps/selvbetjening-ui/sanity/schemaTypes/inntektsjustering/fellesKomponenterSchemaType.ts
