+++ apps/selvbetjening-ui/sanity/tsconfig.json
